+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 # Axon Framework plugin Changelog
 
+## [0.9.3]
+- Remove build-until from plugin.xml, as we generally stay compatible with IDEA
+
 ## [0.9.2]
 - No longer show the routing key inspection warning when the method is not annotated with a relevant handler annotation.
 - Fix false positive on versions not equal to 4